DroneAgent/requirements.txt
tweepy==4.14.0
PyQt5==5.15.10
requests==2.31.0
python-dotenv==1.0.0
PyYAML==6.0.1
pytz==2023.3
matplotlib==3.7.2
Pillow==10.0.0
schedule==1.2.0
openai==1.3.5
google-generativeai==0.3.2
anthropic==0.7.7
